Hotfix unsound Color Prover
The Prover could send random numbers to the Verifier instead of real colorings.
By committing to the coloring beforehand this is avoided. The Verifier will catch the Prover cheating.
If there is a wrongly-colored edge, the Verifier has a chance of challenging the Prover with it.
If the Prover sends the correct opening, the Verifier will see the wrong coloring.
If the Prover sends an opening to different, correct coloring, the Verifier will detect the difference.