Add generator of critical chi-square values
Chi-square requires scipy, which I don't want to add as dependency.
We made do with statically computed values that live in stats.py.
Users can reproduce the results with this script and adjust the
significance level.
This commit also adds a `withScipy` argument to the nix shell which
enables running the script. Pip users have to run pip install scipy
in the virtual environment (todo: update README).