EC: Add ec.static
Optimized EC operations with static lookups.

Points in this module are basically wrappers for integers.
Curves are defined by their discrete log and cannot lie off the curve.
This limits the use cases for this type, but IDK of a proof system off
the top of my head that uses non-curve points.

The points are faster and the randomness is better.

Also switches the util module over to static.