Sudoku: Swap rows and columns
Board.rows is an array of rows that takes a row index.
Each row is an array of cells that takes a column index.

I was confused by x and y vs rows and columns.
The inversion didn't make a difference because sudoku is
symmetric, except for the presets which were previously mirrored.