Sudoku: Add game
The first truly interactive proof in this repo!
It was a lot of work, but also a lot of fun.
We could turn all proofs in this repo into games.
For some more theoretical problems I feel this
wouldn't be a good fit, but the existing graph
problems could be adapted.