EC: Simplify multiplication
Scalars can be multiplied with both Scalars and Points.
This made the implementation more complex and slower.

It is much cleaner to say:
If the first argument of a multiplication is a Scalar, then it
expects a Scalar as second argument and it will produce a Scalar.
If the first argument is a Point, then it expects a Scalar as second
argument and it will produce a Point.

The downside of this change is that we have to write Point * Scalar from
now on, which initially feels weird. I could have implemented everything
using __rmul__, restoring Scalar * Point, but this didn't feel right.
Instead, we will use the convention Point * Scalar from now on.

This will break all code outside local that uses multiplication :)