EC: Add (batch) serialization and batch verification
Points and Scalars can be serialized as integers.
The optional compact parameter can be used to artificually lower the
total number of points. This increases the number of duplicate
transcripts and fills more buckets during the zero-knowledge test.
We choose to compress Points / Scalars instead of the transcript itself
because compressing across value borders could bias the resulting
compressed distribution. If we compress Points / Scalars, we retain a
high probability that similar real / fake distributions pass the test
while different distributions fail the test.

We don't implement (batch) serialization for EC Core because we are
lazy :)