IP: Revamp completeness, soundness, zero-knowledge
We construct a factoring proof to show various properties that an
interactive proof should satisfy.
The explanations don't include how one formally proves these properties,
as we don't formally prove anything in this workshop. We rely on
empirical arguments, which should be intuitive given the explanations in
the IP chapter.
TODO: Explain the zero-knowledge property from the principle of
computational knowledge.