Stats: Update comparison plot
A scatter of plot of both distributions gets messy for a large
number of sample forms (x-axis). It is hard to see how the distributions
differ for a particular x-value because there are many adjacent
x-values. A bar plot is also messy because the bars overlap.

Our solution is to plot one distribution in descending order of counts.
The scatter plot forms a falling line. We plot the other distribution
against the first one. It is easy to see the difference between both
distributions because the viewer can match the shape of the first
distribution onto the dots of the second one.

We move the plot of the absolute differences into a separate subplot
for better overview.