Nix: Support default systems
Sage is not supported by MacOS, so the sage shell is not usable on Mac.
In the future, we could try to make the sage shell available only on
supported systems which would give Mac users a clue that they cannot use
it on their system.

Also update the flake lock file.